chg: normalizacion catalogos y mejora modelo empresa

+++ backend/alembic_tenant/versions/85e51fb13be5_add_factura_detalles.py
@@ -20,6 +20,8 @@ depends_on: Union[str, Sequence[str], None] = None
 
 def upgrade():
     tenant_schema = os.environ.get("TENANT_SCHEMA", "public")
+    if tenant_schema == 'system':
+        return
     op.create_table(
         'factura_detalles',
         sa.Column('id', sa.UUID(as_uuid=True), primary_key=True, server_default=sa.text('gen_random_uuid()')),
@@ -32,6 +34,8 @@ def upgrade():
         schema=tenant_schema
     )
 
-def downgrade():
+def downgrade():    
     tenant_schema = os.environ.get("TENANT_SCHEMA", "public")
+    if tenant_schema == 'system':
+        return
     op.drop_table('factura_detalles', schema=tenant_schema)
