+++ backend/alembic_tenant/versions/90dab873f830_add_partidas.py
@@ -21,6 +21,8 @@ depends_on: Union[str, Sequence[str], None] = None
 
 def upgrade():
     tenant_schema = os.environ.get("TENANT_SCHEMA", "public")
+    if tenant_schema == 'system':
+        return
     op.create_table(
         'partidas',
         sa.Column('id', UUID(as_uuid=True), primary_key=True, server_default=sa.text('gen_random_uuid()')),
@@ -47,5 +49,7 @@ def upgrade():
 
 def downgrade():
     tenant_schema = os.environ.get("TENANT_SCHEMA", "public")
+    if tenant_schema == 'system':
+        return
     op.drop_table('detalle_partidas', schema=tenant_schema)
     op.drop_table('partidas', schema=tenant_schema)
